+++ src/ft_readline/ft_readline_core.c
@@ -61,7 +61,7 @@ static void	register_special_controls(void)
 	register_special_key(CTRL_PAGE_DOWN_KEY, semicolon_five_tilde_key);
 }
 
-t_readline	*aaaaaaaaaaaaaaaaaaa(void)
+t_readline	*get_readline_core(void)
 {
 	static t_readline	*core = NULL;
