+++ src/ft_readline/stdin.c
@@ -6,18 +6,22 @@
 /*   By: lroussel <lroussel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/26 15:16:08 by lroussel          #+#    #+#             */
-/*   Updated: 2025/03/14 19:28:28 by lroussel         ###   ########.fr       */
+/*   Updated: 2025/04/06 13:24:00 by lroussel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "readline.h"
 
-void	read_stdin_keys(char *buffer)
+int	read_stdin_keys(char *buffer)
 {
 	int		byte_read;
 
 	byte_read = read(STDIN_FILENO, buffer, 4095);
 	if (byte_read == -1)
-		exit(4);
+	{
+		buffer[0] = '\0';
+		return (0);
+	}
 	buffer[byte_read] = '\0';
+	return (buffer[0] != '\0');
 }
